+++ externals/coda-oss/modules/c++/include/TestCase.h
@@ -36,10 +36,15 @@
 
 #  define IS_NAN(X) X != X
 
-namespace coda_oss { namespace test {
+namespace test {
 inline void diePrintf(const char* format, const std::string& testName, const char* s)
 {
     sys::diePrintf(format, testName.c_str(), s);
+}
+template<typename TX>
+inline void diePrintf(const char* format, const std::string& testName, const TX& X)
+{
+    diePrintf(format, testName, str::toString(X).c_str());
 }
 
 inline void diePrintf(const char* format, const std::string& testName, const char* file, const char* func, int line)
@@ -47,13 +52,54 @@ inline void diePrintf(const char* format, const std::string& testName, const cha
     sys::diePrintf(format, testName.c_str(), file, func, line);
 }
 
+template<typename TX>
+inline void diePrintf(const char* format, const std::string& testName, const char* file, const char* func, int line,
+    const TX& X)
+{
+    sys::diePrintf(format, testName.c_str(), file, func, line,
+        str::toString(X).c_str());
+}
 template<typename TX1, typename TX2>
 inline void diePrintf(const char* format, const std::string& testName, const char* file, const char* func, int line,
-    TX1&& X1, TX2&& X2)
+    const TX1& X1, const TX2& X2)
 {
     sys::diePrintf(format, testName.c_str(), file, func, line,
         str::toString(X1).c_str(), str::toString(X2).c_str());
 }
+
+template<typename TX1, typename TX2>
+inline void diePrintf_eq(const std::string& testName, const char* file, const char* func, int line,
+    const TX1& X1, const TX2& X2)
+{
+    diePrintf("%s (%s,%s,%d): FAILED: Recv'd %s, Expected %s\n", testName, file, func, line, X1, X2);      
+}
+#define CODA_OSS_test_diePrintf_eq_(X1, X2) test::diePrintf_eq(testName, __FILE__, SYS_FUNC, __LINE__, X1, X2)
+
+template<typename TX1, typename TX2, typename TMsg>
+inline void diePrintf_eq_msg(const std::string& testName, const TMsg& msg, const char* file, int line,
+    const TX1& X1, const TX2& X2)
+{
+    sys::diePrintf("%s (%s,%d): FAILED (%s): Recv'd %s, Expected %s\n", testName.c_str(), file, line, (msg).c_str(),
+        str::toString((X1)).c_str(), str::toString((X2)).c_str());       
+}
+#define CODA_OSS_test_diePrintf_eq_msg_(msg, X1, X2) test::diePrintf_eq_msg(testName, msg, __FILE__, __LINE__, X1, X2)
+
+template<typename TX1, typename TX2>
+inline void diePrintf_ne(const std::string& testName, const char* file, const char* func, int line,
+    const TX1& X1, const TX2& X2)
+{
+    diePrintf("%s (%s,%s,%d): FAILED: Recv'd %s should not equal %s\n", testName, file, func, line, X1, X2);
+}
+#define CODA_OSS_test_diePrintf_not_eq_(X1, X2) test::diePrintf_ne(testName, __FILE__, SYS_FUNC, __LINE__, X1, X2)
+
+template<typename TX1, typename TX2, typename TMsg>
+inline void diePrintf_ne_msg(const std::string& testName, const TMsg& msg, const char* file, int line,
+    const TX1& X1, const TX2& X2)
+{
+    sys::diePrintf("%s (%s,%d): FAILED (%s): Recv'd %s should not equal %s\n", testName.c_str(), file, line, (msg).c_str(),
+        str::toString((X1)).c_str(), str::toString((X2)).c_str());       
+}
+#define CODA_OSS_test_diePrintf_not_eq_msg_(msg, X1, X2) test::diePrintf_ne_msg(testName, msg, __FILE__, __LINE__, X1, X2)
 
 template <typename TFunc>
 inline void check(TFunc f, const std::string& testName)
@@ -65,7 +111,7 @@ inline void check(TFunc f, const std::string& testName)
     }
     catch (const except::Throwable& ex)
     {
-        diePrintf("%s: FAILED: Exception thrown: %s\n", testName, ex.toString().c_str());
+        diePrintf("%s: FAILED: Exception thrown: %s\n", testName, ex.toString());
     }
     catch (const except::Throwable11& ex)
     {
@@ -74,7 +120,7 @@ inline void check(TFunc f, const std::string& testName)
 }
 
 template<typename TX>
-inline void assert_(TX&& X, const std::string& testName, const char* file, const char* func, int line)
+inline void assert_(const TX& X, const std::string& testName, const char* file, const char* func, int line)
 {
     if (!X)
     {
@@ -83,16 +129,24 @@ inline void assert_(TX&& X, const std::string& testName, const char* file, const
 }
 
 template<typename TX>
-inline void assert_null(TX&& X, const std::string& testName,  const char* file, const char* func, int line)
+inline void assert_null(const TX& X, const std::string& testName,  const char* file, const char* func, int line)
 {
     if ((X != nullptr) || (!(X == nullptr)))
     {
         diePrintf("%s (%s,%s,%d): FAILED: Value should be NULL\n", testName, file, func, line);
     }
 }
+template<typename TX>
+inline void assert_not_null(const TX& X, const std::string& testName,  const char* file, const char* func, int line)
+{
+    if ((X == nullptr) || (!(X != nullptr)))
+    {
+        diePrintf("%s (%s,%s,%d): FAILED: Value should *not* be NULL\n", testName, file, func, line);
+    }
+}
 
 template<typename TX>
-inline void test_assert_false(TX&& X, const std::string& testName,  const char* file, const char* func, int line)
+inline void test_assert_false(const TX& X, const std::string& testName,  const char* file, const char* func, int line)
 {
     if (X)
     {
@@ -101,7 +155,7 @@ inline void test_assert_false(TX&& X, const std::string& testName,  const char* 
 }
 
 template<typename TX>
-inline void test_assert_true(TX&& X, const std::string& testName,  const char* file, const char* func, int line)
+inline void test_assert_true(const TX& X, const std::string& testName,  const char* file, const char* func, int line)
 {
     if (!X)
     {
@@ -109,105 +163,37 @@ inline void test_assert_true(TX&& X, const std::string& testName,  const char* f
     }
 }
 
-// These cause warnings (or even errors, depending on compile settings) if the types have different signedness.
-// assert_eq(v.size(), 1, ...);
-/*
-template<typename TX1, typename TX2>
-inline void assert_eq(TX1&& X1, TX2&& X2,
-    const std::string& testName,  const char* file, const char* func, int line)
-{
-    if ((X1 != X2) || (!(X1 == X2)))
-    {
-        diePrintf("%s (%s,%s,%d): FAILED: Recv'd %s, Expected %s\n", testName, file, func, line, X1, X2);
-    }
-}
-template<typename TX1, typename TX2>
-inline void assert_eq_msg(const std::string& msg, TX1&& X1, TX2&& X2,
-    const std::string& testName,  const char* file, int line)
-{
-    if ((X1 != X2) || (!(X1 == X2)))
-    {
-        die_printf("%s (%s,%d): FAILED (%s): Recv'd %s, Expected %s\n", testName.c_str(), file, line, 
-                   msg.c_str(), str::toString(X1).c_str(), str::toString(X2).c_str());
-    }
-}
+// You might be tempted to write something like
+//   template<typename TX1, typename TX2>
+//  inline void assert_eq(const TX1& X1, const TX2& X2, ...) {}
+// A nice thought, but the compiler is able to do more implicit conversions when the code is "inline"
+//    if (!(X1 == X2)) { ... }
+// behaves differently when it is in the functon (X1 and X2 are arguments).  An easy example
+// is std::vector::size() (size_t) compared to a literal 1 which is an "int" not "size_t".
 
-template<typename TX1, typename TX2>
-inline void assert_not_eq(TX1&& X1, TX2&& X2,
-    const std::string& testName, const char* file, const char* func, int line)
-{
-    if ((X1 == X2) || (!(X1 != X2)))
-    {
-        diePrintf("%s (%s,%s,%d): FAILED: Recv'd %s should not equal %s\n", testName, file, func, line, X1, X2);
-    }
-}
-*/
 template<typename TX1, typename TX2, typename TEPS>
-inline void assert_almost_eq_eps(TX1&& X1, TX2&& X2, TEPS&& EPS,
+inline void assert_almost_eq_eps(const TX1& X1, const TX2& X2, const TEPS& EPS,
     const std::string& testName, const char* file, const char* func, int line)
 {
     const auto abs_difference = std::abs(X1 - X2);
     if (abs_difference > EPS || IS_NAN(abs_difference))
     {
-        diePrintf("%s (%s,%d): FAILED: Recv'd %s, Expected %s\n", testName, file, func, line, X1, X2);
-    }
-}
-template<typename TX1, typename TX2>
-inline void assert_almost_eq(TX1&& X1, TX2&& X2,
-    const std::string& testName,  const char* file, const char* func, int line)
-{
-    const auto abs_difference = std::abs(X1 - X2);
-    if (abs_difference > std::numeric_limits<float>::epsilon() || IS_NAN(abs_difference))
-    {
         diePrintf("%s (%s,%s,%d): FAILED: Recv'd %s, Expected %s\n", testName, file, func, line, X1, X2);
     }
 }
-/*
 template<typename TX1, typename TX2>
-inline void assert_greater_eq(TX1&& X1, TX2&& X2,
+inline void assert_almost_eq(const TX1& X1, const TX2& X2,
     const std::string& testName,  const char* file, const char* func, int line)
 {
-    if ((X1 < X2) || (!(X1 >= X2)))
-    {
-        diePrintf("%s (%s,%s,%d): FAILED: Value should be greater than or equal\n", testName, file, func, line);
-    }
+    const auto eps = std::numeric_limits<float>::epsilon();
+    assert_almost_eq_eps(X1, X2, eps, testName, file, func, line);
 }
 
-template<typename TX1, typename TX2>
-inline void assert_greater(TX1&& X1, TX2&& X2,
-    const std::string& testName,  const char* file, const char* func, int line)
-{
-    if ((X1 <= X2) || (!(X1 > X2)))
-    {
-        diePrintf("%s (%s,%s,%d): FAILED: Value should be greater than\n", testName, file, func, line);
-    }
-}
-
-template<typename TX1, typename TX2>
-inline void assert_lesser_eq(TX1&& X1, TX2&& X2,
-    const std::string& testName,  const char* file, const char* func, int line)
-{
-    if ((X1 > X2) || (!(X1 <= X2)))
-    {
-        diePrintf("%s (%s,%s,%d): FAILED: Value should be less than or equal\n", testName, file, func, line);
-    }
-}
-
-template<typename TX1, typename TX2>
-inline void assert_lesser(TX1&& X1, TX2&& X2,
-    const std::string& testName,  const char* file, const char* func, int line)
-{
-    if ((X1 >= X2) || (!(X1 < X2)))
-    {
-        diePrintf("%s (%s,%s,%d): FAILED: Value should be less than\n", testName, file, func, line);
-    }
-}
-*/
 template<typename Tmsg>
-inline void fail(Tmsg&& msg,
+inline void fail(const Tmsg& msg,
     const std::string& testName,  const char* file, const char* func, int line)
 {
-    die_printf("%s (%s,%s,%d): FAILED: %s\n", testName.c_str(), file, func, line, str::toString(msg).c_str());
+    diePrintf("%s (%s,%s,%d): FAILED: %s\n", testName, file, func, line, msg);
 }
 
 template<typename TFunc>
@@ -277,32 +263,23 @@ inline int main(TFunc f)
     }
     return EXIT_FAILURE;
 }
-}}
-#define TEST_CHECK(X) coda_oss::test::check([&](const std::string& testName) { X(testName); }, #X)
-#define TEST_ASSERT(X) coda_oss::test::assert_(X, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_ASSERT_NULL(X) coda_oss::test::assert_null(X, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_ASSERT_FALSE(X) coda_oss::test::test_assert_false(X, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_ASSERT_TRUE(X) coda_oss::test::test_assert_true(X, testName, __FILE__, SYS_FUNC, __LINE__)
-/*
-#define TEST_ASSERT_EQ(X1, X2) coda_oss::test::assert_eq(X1, X2, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_ASSERT_EQ_MSG(msg, X1, X2) coda_oss::test::assert_eq_msg(msg, X1, X2, testName, __FILE__, __LINE__)
-#define TEST_ASSERT_NOT_EQ(X1, X2) coda_oss::test::assert_not_eq(X1, X2, testName, __FILE__, SYS_FUNC, __LINE__)
-*/
-#define TEST_ASSERT_ALMOST_EQ_EPS(X1, X2, EPS) coda_oss::test::assert_almost_eq_eps(X1, X2, EPS, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_ASSERT_ALMOST_EQ(X1, X2) coda_oss::test::assert_almost_eq(X1, X2, testName, __FILE__, SYS_FUNC, __LINE__)
-/*
-#define TEST_ASSERT_GREATER_EQ(X1, X2) coda_oss::test::assert_greater_eq(X1, X2, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_ASSERT_GREATER(X1, X2) coda_oss::test::assert_greater(X1, X2, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_ASSERT_LESSER_EQ(X1, X2) coda_oss::test::assert_lesser_eq(X1, X2, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_ASSERT_LESSER(X1, X2) coda_oss::test::assert_lesser(X1, X2, testName, __FILE__, SYS_FUNC, __LINE__)
-*/
-#define TEST_FAIL(msg) coda_oss::test::fail(msg, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_EXCEPTION(X) coda_oss::test::exception([&](){(X);}, testName, __FILE__, SYS_FUNC, __LINE__)
-#define TEST_THROWS(X) coda_oss::test::throws([&](){(X);}, testName, __FILE__, SYS_FUNC, __LINE__)
-# define TEST_SPECIFIC_EXCEPTION(X, Y) coda_oss::test::specific_exception<Y>([&](){(X);}, \
+}
+#define TEST_CHECK(X) test::check([&](const std::string& testName) { X(testName); }, #X)
+#define TEST_ASSERT(X) test::assert_(X, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_ASSERT_NULL(X) test::assert_null(X, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_ASSERT_NOT_NULL(X) test::assert_not_null(X, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_ASSERT_FALSE(X) test::test_assert_false(X, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_ASSERT_TRUE(X) test::test_assert_true(X, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_ASSERT_ALMOST_EQ_EPS(X1, X2, EPS) test::assert_almost_eq_eps(X1, X2, EPS, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_ASSERT_ALMOST_EQ(X1, X2) test::assert_almost_eq(X1, X2, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_FAIL(msg) test::fail(msg, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_EXCEPTION(X) test::exception([&](){(X);}, testName, __FILE__, SYS_FUNC, __LINE__)
+#define TEST_THROWS(X) test::throws([&](){(X);}, testName, __FILE__, SYS_FUNC, __LINE__)
+# define TEST_SPECIFIC_EXCEPTION(X, Y) test::specific_exception<Y>([&](){(X);}, \
     "%s (%s,%s,%d): FAILED: Should have thrown exception: " # Y ,  testName, __FILE__, SYS_FUNC, __LINE__)
 #  define TEST_CASE(X) void X(std::string testName)
-#define TEST_MAIN(X) int main(int argc, char** argv) { return coda_oss::test::main([&](){X;}); }
+#define TEST_MAIN(X) int main() { return test::main([&](){X;}); }
+#define TEST_MAIN_ARGS(X) int main(int argc, char* argv[]) { return test::main([&](){X;}); }
 /*
 #  define TEST_CHECK(X) try{ X(std::string(#X)); std::cerr << #X << ": PASSED" << std::endl; } \
     catch(const except::Throwable& ex) { die_printf("%s: FAILED: Exception thrown: %s\n", std::string(#X).c_str(), ex.toString().c_str()); } \
@@ -312,18 +289,22 @@ inline int main(TFunc f)
 #  define TEST_ASSERT_FALSE(X) if ((X)) { die_printf("%s (%s,%s,%d): FAILED: Value should evaluate to false\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__); }
 #  define TEST_ASSERT_TRUE(X) if (!(X)) { die_printf("%s (%s,%s,%d): FAILED: Value should evaluate to true\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__); }
 */
-#  define TEST_ASSERT_EQ(X1, X2) if ((X1) != (X2)) { die_printf("%s (%s,%s,%d): FAILED: Recv'd %s, Expected %s\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__, str::toString(X1).c_str(), str::toString(X2).c_str()); }
-#  define TEST_ASSERT_EQ_MSG(msg, X1, X2) if ((X1) != (X2)) die_printf("%s (%s,%d): FAILED (%s): Recv'd %s, Expected %s\n", testName.c_str(), __FILE__, __LINE__, (msg).c_str(), str::toString((X1)).c_str(), str::toString((X2)).c_str());
-#  define TEST_ASSERT_NOT_EQ(X1, X2) if ((X1) == (X2)) { die_printf("%s (%s,%s,%d): FAILED: Recv'd %s should not equal %s\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__, str::toString(X1).c_str(), str::toString(X2).c_str()); }
-#  define TEST_ASSERT_NOT_EQ_MSG(msg, X1, X2) if ((X1) == (X2)) die_printf("%s (%s,%d): FAILED (%s): Recv'd %s should not equal %s\n", testName.c_str(), __FILE__, __LINE__, (msg).c_str(), str::toString((X1)).c_str(), str::toString((X2)).c_str());
-/*
-#  define TEST_ASSERT_ALMOST_EQ_EPS(X1, X2, EPS) if (std::abs((X1) - (X2)) > EPS || IS_NAN(std::abs((X1) - (X2)))) die_printf("%s (%s,%d): FAILED: Recv'd %s, Expected %s\n", testName.c_str(), __FILE__, __LINE__, str::toString((X1)).c_str(), str::toString((X2)).c_str());
-#  define TEST_ASSERT_ALMOST_EQ(X1, X2) if (std::abs((X1) - (X2)) > std::numeric_limits<float>::epsilon() || IS_NAN(std::abs((X1) - (X2)))) { die_printf("%s (%s,%s,%d): FAILED: Recv'd %s, Expected %s\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__, str::toString(X1).c_str(), str::toString(X2).c_str()); }
-*/
-#  define TEST_ASSERT_GREATER_EQ(X1, X2) if ((X1) < X2) { die_printf("%s (%s,%s,%d): FAILED: Value should be greater than or equal\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__); }
-#  define TEST_ASSERT_GREATER(X1, X2) if ((X1) <= X2) { die_printf("%s (%s,%s,%d): FAILED: Value should be greater than\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__); }
-#  define TEST_ASSERT_LESSER_EQ(X1, X2) if ((X1) > X2) { die_printf("%s (%s,%s,%d): FAILED: Value should be less than or equal\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__); }
-#  define TEST_ASSERT_LESSER(X1, X2) if ((X1) >= X2) { die_printf("%s (%s,%s,%d): FAILED: Value should be less than\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__); }
+//#define CODA_OSS_test_eq_(X1, X2) (((X1) == (X2)) && ((X2) == (X1))) // X1 == X2 means X2 == X1
+#define CODA_OSS_test_eq_(X1, X2) ((X1) == (X2)) // above breaks CODA :-(
+//#define CODA_OSS_test_ne_(X1, X2) (((X1) != (X2)) && ((X2) != (X1))) // X1 != X2 means X2 != X1
+#define CODA_OSS_test_ne_(X1, X2) ((X1) != (X2)) // above breaks CODA :-(
+//#define CODA_OSS_test_ne(X1, X2) (CODA_OSS_test_ne_(X1, X2) && !CODA_OSS_test_eq_(X1, X2))
+#define CODA_OSS_test_ne(X1, X2) CODA_OSS_test_ne_(X1, X2) // above breaks CODA :-(
+#define TEST_ASSERT_EQ(X1, X2) if (CODA_OSS_test_ne((X1), (X2))) { CODA_OSS_test_diePrintf_eq_(X1, X2); }
+#define TEST_ASSERT_EQ_MSG(msg, X1, X2) if (CODA_OSS_test_ne((X1), (X2))) { CODA_OSS_test_diePrintf_eq_msg_(msg, X1, X2); }
+//#define CODA_OSS_test_eq(X1, X2) (CODA_OSS_test_eq_(X1, X2) && !CODA_OSS_test_ne_(X1, X2))
+#define CODA_OSS_test_eq(X1, X2) CODA_OSS_test_eq_(X1, X2) // above breaks CODA :-(
+#define TEST_ASSERT_NOT_EQ(X1, X2) if (CODA_OSS_test_eq((X1), (X2))) { CODA_OSS_test_diePrintf_not_eq_(X1, X2); }
+#define TEST_ASSERT_NOT_EQ_MSG(msg, X1, X2) if (CODA_OSS_test_eq((X1), (X2))) { CODA_OSS_test_diePrintf_not_eq_msg_(msg, X1, X2); }
+#  define TEST_ASSERT_GREATER_EQ(X1, X2) if ((X1) < X2) { test::diePrintf("%s (%s,%s,%d): FAILED: Value should be greater than or equal\n", testName, __FILE__, SYS_FUNC, __LINE__); }
+#  define TEST_ASSERT_GREATER(X1, X2) if ((X1) <= X2) { test::diePrintf("%s (%s,%s,%d): FAILED: Value should be greater than\n", testName, __FILE__, SYS_FUNC, __LINE__); }
+#  define TEST_ASSERT_LESSER_EQ(X1, X2) if ((X1) > X2) { test::diePrintf("%s (%s,%s,%d): FAILED: Value should be less than or equal\n", testName, __FILE__, SYS_FUNC, __LINE__); }
+#  define TEST_ASSERT_LESSER(X1, X2) if ((X1) >= X2) { test::diePrintf("%s (%s,%s,%d): FAILED: Value should be less than\n", testName, __FILE__, SYS_FUNC, __LINE__); }
 /*
 #  define TEST_FAIL(msg) die_printf("%s (%s,%s,%d): FAILED: %s\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__, str::toString(msg).c_str());
 #  define TEST_EXCEPTION(X) try{ (X); die_printf("%s (%s,%s,%d): FAILED: Should have thrown exception\n", testName.c_str(), __FILE__, SYS_FUNC, __LINE__); } catch (const except::Throwable&){} catch (const except::Throwable11&){}