+++ externals/coda-oss/modules/c++/math/include/import/math.h
@@ -1,8 +1,8 @@
 /* =========================================================================
- * This file is part of mem-c++
+ * This file is part of math-c++
  * =========================================================================
  *
- * (C) Copyright 2004 - 2018, MDA Information Systems LLC
+ * (C) Copyright 2004 - 2014, MDA Information Systems LLC
  *
  * mem-c++ is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -20,30 +20,14 @@
  *
  */
 
-#ifndef CODA_OSS_mem_SharedPtr_h_INCLUDED_
-#define CODA_OSS_mem_SharedPtr_h_INCLUDED_
+#ifndef CODA_OSS_math_import_h_INCLUDED_
+#define CODA_OSS_math_import_h_INCLUDED_
 #pragma once
 
-#include <memory>
+#include <math/Bessel.h>
+#include <math/Constants.h>
+#include <math/ConvexHull.h>
+#include <math/Round.h>
+#include <math/Utilities.h>
 
-// Adding a forward declaration here in order to have SWIG correctly generate
-// bindings for SharedPtr.
-//
-// SWIG's default behavior is to not process includes recursively when given an
-// %include directive (neither SharedPtrCpp11 or SharedPtrLegacy will get
-// processed) - including this header without the forward declaration
-// as part of a SWIG interface will result in a missing template type error.
-namespace mem
-{
-#if !CODA_OSS_enable_mem_SharedPtr
-template <typename T>
-using SharedPtr = std::shared_ptr<T>;
-#else
-template<typename T>
-class SharedPtr;
-#endif
-}
-
-#include <mem/SharedPtrCpp11.h>
-
-#endif // CODA_OSS_mem_SharedPtr_h_INCLUDED_
+#endif  // CODA_OSS_math_import_h_INCLUDED_