+++ externals/coda-oss/modules/c++/str/include/str/EncodedString.h
@@ -5,7 +5,7 @@
  * (C) Copyright 2004 - 2014, MDA Information Systems LLC
  * (C) Copyright 2022, Maxar Technologies, Inc.
  *
- * xml.lite-c++ is free software; you can redistribute it and/or modify
+ * str-c++ is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
  * the Free Software Foundation; either version 3 of the License, or
  * (at your option) any later version.
@@ -28,6 +28,7 @@
 #include <string>
 #include <ostream>
 
+#include "config/Exports.h"
  #include "str/EncodedStringView.h"
 
 /*!
@@ -45,7 +46,7 @@
 
 namespace str
 {
-class EncodedString final
+class CODA_OSS_API EncodedString final
 {
     std::string s_;
     // We can do most everything through the view, so keep one around.
@@ -66,19 +67,17 @@ public:
     ~EncodedString() = default;
     EncodedString(const EncodedString&);
     EncodedString& operator=(const EncodedString&);
-    EncodedString(EncodedString&&) noexcept;
-    EncodedString& operator=(EncodedString&&) noexcept;
+    EncodedString(EncodedString&&);
+    EncodedString& operator=(EncodedString&&);
 
-    explicit EncodedString(const coda_oss::u8string& s);
     explicit EncodedString(coda_oss::u8string::const_pointer);
+    explicit EncodedString(const coda_oss::u8string& s);
     explicit EncodedString(const str::W1252string&);
     explicit EncodedString(str::W1252string::const_pointer);
     explicit EncodedString(const std::string&);  // Assume platform native encoding: UTF-8 on Linux, Windows-1252 on Windows
     explicit EncodedString(std::string::const_pointer);  // Assume platform native encoding: UTF-8 on Linux, Windows-1252 on Windows
     explicit EncodedString(const std::u16string&); // converted to UTF-8 for storage
-    //explicit EncodedString(std::u16string::const_pointer); // no wcslen() for std::u16string::value_type
     explicit EncodedString(const std::u32string&); // converted to UTF-8 for storage
-    //explicit EncodedString(std::u32string::const_pointer); // no wcslen() for std::u32string::value_type
     explicit EncodedString(const std::wstring&);  // Assume platform native encoding: UTF-32 on Linux, UTF-16 on Windows
     explicit EncodedString(std::wstring::const_pointer);  // can call wcslen()
 
@@ -87,10 +86,8 @@ public:
     EncodedString& operator=(const EncodedStringView&);
     
     // Input is encoded as specified on all platforms.
-    static EncodedString fromUtf8(const std::string&);
-    static EncodedString fromWindows1252(const std::string&);
-    static EncodedString fromUtf16(const std::wstring&); // not currently implemetned, no need
-    static EncodedString fromUtf32(const std::wstring&); // not currently implemetned, no need
+    //static EncodedString fromUtf16(const std::wstring&); // not currently implemetned, no need
+    //static EncodedString fromUtf32(const std::wstring&); // not currently implemetned, no need
     
     // For "complex" operatations, use the view.
     const EncodedStringView& view() const