+++ six/modules/c++/six/unittests/test_serialize.cpp
@@ -139,7 +139,7 @@ TEST_CASE(VectorSerialize)
     TEST_ASSERT_TRUE(testVector<double>(length, true));
 }
 
-TEST_MAIN((void)argv; (void)argc;
+TEST_MAIN(
     srand(static_cast<unsigned int>(time(NULL)));
     TEST_CHECK(ScalarSerialize);
     TEST_CHECK(VectorSerialize);